add interface in types folder
