+++ src/types/index.ts
@@ -11,7 +11,25 @@ interface HeaderKey {
   headerName: string;
 }
 
+interface Data {
+  [key: string]: number | string | boolean | null;
+}
+type Direction = 'asc' | 'desc';
+type Operator = 'is' | 'is not' | 'is empty' | 'is not empty';
+
+interface ObjectUnion {
+  [key: string]: any;
+}
+interface TooltipProps {
+  content: string;
+  children: React.ReactElement;
+}
 export {
   HeaderKey,
   GlobalStateContextVals,
+  Data,
+  Direction,
+  Operator,
+  ObjectUnion,
+  TooltipProps,
 };
